+++ framework/widgets/ActiveField.php
@@ -261,15 +261,16 @@ class ActiveField extends Component
      * @param array|null $options the tag options in terms of name-value pairs. It will be merged with [[labelOptions]].
      * The options will be rendered as the attributes of the resulting tag. The values will be HTML-encoded using
      * [[Html::encode()]]. If a value is `null`, the corresponding attribute will not be rendered.
+     *
      * The following special options are recognized:
-     * - `label`: string|false|null, if specified in $options (or previously set in [[labelOptions]]), this value will
-     *   be used when the $label parameter is `null`. If set to `false`, the label will not be rendered (same behavior
-     *   as passing $label = `false`). The $label parameter always takes precedence over this option.
-     * - `tag`: string|false, specifies the tag name for the label element.
-     *   - If not specified, defaults to `'label'`.
-     *   - If set to `false`, the label content will be rendered as raw HTML without any wrapper tag.
-     *   - If set to `'label'`, uses [[Html::activeLabel()]] to generate a standard label element.
-     *   - If set to another tag name (for example, `'span'`, `'div'`), uses [[Html::tag()]] to generate that element.
+     *
+     * - `label`: `string|false|null`, the label text. Used as fallback when `$label` parameter is `null`.
+     *   If set to `false`, the label will not be rendered. The `$label` parameter always takes precedence over this
+     *   option.
+     * - `tag`: `string|false`, the tag name for the label element. Defaults to `'label'`.
+     *   If set to `false`, the label content is rendered without any wrapping tag.
+     *   If set to `'label'`, [[Html::activeLabel()]] is used (default behavior).
+     *   Any other string value (e.g. `'span'`, `'div'`, `'h3'`) will use [[Html::tag()]] with that element.
      *
      * @return $this the field object itself.
      */
@@ -548,21 +549,13 @@ class ActiveField extends Component
      *   it will take the default value `0`. This method will render a hidden input so that if the radio button
      *   is not checked and is submitted, the value of this attribute will still be submitted to the server
      *   via the hidden input. If you do not want any hidden input, you should explicitly set this option as `null`.
-     * - `label`: string|false|null, a label displayed next to the radio.
-     *   - If a string, it will NOT be HTML-encoded. Therefore you can pass in HTML code such as an image tag.
-     *   - If this is coming from end users, you should [[Html::encode()|encode]] it to prevent XSS attacks.
-     *   - If `false`, no label will be rendered.
-     *   - If `null`, no label will be rendered (same as `false`).
-     *   When this option is specified as a string and `enclosedByLabel` is `true`, the radio will be enclosed by a
-     *   label tag.
-     * - `labelOptions`: array, the HTML attributes for the label tag. This is only used when the `label` option is
-     *   specified and `enclosedByLabel` is `false`. The following special option is recognized:
-     *   - `tag`: string|false, specifies the tag name for the label element.
-     *     - If `labelOptions` is not provided, the label content will be rendered as raw HTML without any wrapper tag.
-     *     - If `labelOptions` is provided but `tag` is not specified, defaults to `'label'`.
-     *     - If set to `'label'`, uses [[Html::activeLabel()]] to generate a standard label element.
-     *     - If set to another tag name (for example, `'span'`, `'div'`), uses [[Html::tag()]] to generate that element.
-     *     - If set to `false`, the label content will be rendered as raw HTML without any wrapper tag.
+     * - `label`: `string|false|null`, a label displayed next to the radio button. It will NOT be HTML-encoded.
+     *   Therefore you can pass in HTML code such as an image tag. If this is coming from end users,
+     *   you should [[Html::encode()|encode]] it to prevent XSS attacks.
+     *   When this option is specified, the radio button will be enclosed by a label tag.
+     *   If you do not want any label, you should explicitly set this option as `null`.
+     * - `labelOptions`: array, the HTML attributes for the label tag. This is only used when the `label` option
+     *   is specified. Supports a `tag` sub-option (`'label'`/`false`/custom tag) for flexible label rendering.
      *
      * The rest of the options will be rendered as the attributes of the resulting tag. The values will
      * be HTML-encoded using [[Html::encode()]]. If a value is `null`, the corresponding attribute will not be rendered.
@@ -583,11 +576,11 @@ class ActiveField extends Component
         $this->addAriaAttributes($options);
         $this->adjustLabelFor($options);
 
-        if ($enclosedByLabel) {
-            $this->parts['{label}'] = '';
-        } else {
+        if (!$enclosedByLabel) {
             $options = $this->generateLabel($options);
         }
+
+        $this->parts['{label}'] ??= '';
 
         $this->parts['{input}'] = Html::activeRadio($this->model, $this->attribute, $options);
 
@@ -603,21 +596,13 @@ class ActiveField extends Component
      *   it will take the default value `0`. This method will render a hidden input so that if the checkbox
      *   is not checked and is submitted, the value of this attribute will still be submitted to the server
      *   via the hidden input. If you do not want any hidden input, you should explicitly set this option as `null`.
-     * - `label`: string|false|null, a label displayed next to the checkbox.
-     *   - If a string, it will NOT be HTML-encoded. Therefore you can pass in HTML code such as an image tag.
-     *   - If this is coming from end users, you should [[Html::encode()|encode]] it to prevent XSS attacks.
-     *   - If `false`, no label will be rendered.
-     *   - If `null`, no label will be rendered (same as `false`).
-     *   When this option is specified as a string and `enclosedByLabel` is `true`, the checkbox will be enclosed by a
-     *   label tag.
+     * - `label`: `string|false|null`, a label displayed next to the checkbox. It will NOT be HTML-encoded.
+     *   Therefore you can pass in HTML code such as an image tag. If this is coming from end users, you should
+     *   [[Html::encode()|encode]] it to prevent XSS attacks.
+     *   When this option is specified, the checkbox will be enclosed by a label tag.
+     *   If you do not want any label, you should explicitly set this option as `null`.
      * - `labelOptions`: array, the HTML attributes for the label tag. This is only used when the `label` option is
-     *   specified and `enclosedByLabel` is `false`. The following special option is recognized:
-     *   - `tag`: string|false, specifies the tag name for the label element.
-     *     - If `labelOptions` is not provided, the label content will be rendered as raw HTML without any wrapper tag.
-     *     - If `labelOptions` is provided but `tag` is not specified, defaults to `'label'`.
-     *     - If set to `'label'`, uses [[Html::activeLabel()]] to generate a standard label element.
-     *     - If set to another tag name (for example, `'span'`, `'div'`), uses [[Html::tag()]] to generate that element.
-     *     - If set to `false`, the label content will be rendered as raw HTML without any wrapper tag.
+     *   specified. Supports a `tag` sub-option (`'label'`/`false`/custom tag) for flexible label rendering.
      *
      * The rest of the options will be rendered as the attributes of the resulting tag. The values will
      * be HTML-encoded using [[Html::encode()]]. If a value is `null`, the corresponding attribute will not be rendered.
@@ -638,11 +623,11 @@ class ActiveField extends Component
         $this->addAriaAttributes($options);
         $this->adjustLabelFor($options);
 
-        if ($enclosedByLabel) {
-            $this->parts['{label}'] = '';
-        } else {
+        if (!$enclosedByLabel) {
             $options = $this->generateLabel($options);
         }
+
+        $this->parts['{label}'] ??= '';
 
         $this->parts['{input}'] = Html::activeCheckbox($this->model, $this->attribute, $options);
 
@@ -945,31 +930,38 @@ class ActiveField extends Component
      */
     protected function generateLabel(array $options): array
     {
-        if (isset($options['label'])) {
-            if ($options['label'] === false) {
-                $this->parts['{label}'] = '';
-            } elseif (($this->parts['{label}'] ?? '') === '') {
-                $tag = false;
+        if (isset($options['label']) && $options['label'] !== false && !array_key_exists('{label}', $this->parts)) {
+            $tag = false;
+            $labelOptions = $this->labelOptions;
 
-                if (!empty($options['labelOptions'])) {
-                    $tag = $options['labelOptions']['tag'] ?? 'label';
+            if (array_key_exists('labelOptions', $options)) {
+                $tag = $options['labelOptions']['tag'] ?? 'label';
 
-                    unset($options['labelOptions']['tag']);
+                unset($options['labelOptions']['tag']);
 
-                    $this->labelOptions = array_merge($this->labelOptions, $options['labelOptions']);
-                }
+                $labelOptions = [
+                    ...$labelOptions,
+                    ...$options['labelOptions'],
+                ];
+            }
 
-                if ($tag === false) {
-                    $this->parts['{label}'] = $options['label'];
-                } elseif ($tag === 'label') {
-                    $this->parts['{label}'] = Html::activeLabel(
-                        $this->model,
-                        $this->attribute,
-                        array_merge(['label' => $options['label']], $this->labelOptions),
-                    );
-                } else {
-                    $this->parts['{label}'] = Html::tag($tag, $options['label'], $this->labelOptions);
-                }
+            if ($tag === false) {
+                $this->parts['{label}'] = $options['label'];
+            } elseif ($tag === 'label') {
+                $this->parts['{label}'] = Html::activeLabel(
+                    $this->model,
+                    $this->attribute,
+                    [
+                        ...$labelOptions,
+                        'label' => $options['label'],
+                    ],
+                );
+            } else {
+                $labelOpts = $labelOptions;
+
+                unset($labelOpts['label'], $labelOpts['for']);
+
+                $this->parts['{label}'] = Html::tag($tag, $options['label'], $labelOpts);
             }
         }